requirements.txt
Django==4.2.6
django-environ==0.11.2
django-ninja==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9